stocktrader/pyproject.toml
[tool.poetry]
name = "stocktrader"
version = "0.1.0"
description = "Application for trading on stock markets"
authors = ["twelve <kana.suzucki@gmail.com>"]
license = "GNU General Public License v3.0"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
pyjwt = "^2.6.0"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
pyex = "^0.5.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest-django = "^4.5.2"


[tool.poetry.group.database.dependencies]
psycopg2-binary = "^2.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
